pyproject.toml
[project]
name = "render-tag"
version = "0.1.0"
description = "Procedural 3D data generation for AprilTag training (Offline)."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Noe Fontana" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.8",
    "pydantic-settings>=2.3",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "huggingface-hub>=0.24.0",
    "orjson>=3.10",
]

[project.optional-dependencies]
# Required for 3D rendering
backend = [
    "blenderproc>=2.8.0",
]
# Required for 2D visualization and Shadow Render
viz = [
    "matplotlib>=3.7",
]
# Required for dataset auditing and analytics
audit = [
    "polars>=1.0.0",
    "plotly>=5.15.0",
]
# Required for distributed orchestration and telemetry
orchestration = [
    "pyzmq>=25.0.0",
    "gputil>=1.4.0",
]
# Required for asset synchronization with Hugging Face
assets = [
]
# Install everything
all = [
    "render-tag[backend,viz,audit,orchestration,assets]",
]

[project.urls]
Repository = "https://github.com/NoeFontana/render-tag"

[project.scripts]
render-tag = "render_tag.cli:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/render_tag"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff",
    "mypy",
    "ty",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
    "pymdown-extensions>=10.20.1",
    "import-linter",
    "fake-bpy-module-4.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["N803", "N806", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["render_tag"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.src]
include = ["src", "tests"]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.rules]
# Standard rules enabled
# unresolved-import = "ignore" # Removed to enable Blender API checking
# possibly-missing-attribute = "ignore"
# unresolved-attribute = "ignore"
invalid-method-override = "ignore"
invalid-assignment = "ignore"
unused-ignore-comment = "ignore"
invalid-argument-type = "ignore"

[tool.pytest.ini_options]
addopts = "-n 4 --dist loadfile"
markers = [
    "integration: marks tests as slow integration tests (deselect with '-m \"not integration\"')",
    "timeout: marks tests with a timeout duration",
]
testpaths = ["tests"]

[tool.importlinter]
root_package = "render_tag"

["importlinter:contract:1"]
name = "Backend must not import host logic"
type = "layers"
containers = ["render_tag"]
layers = [
    "cli",
    "orchestration",
    "generation",
    "core",
    "backend",
]
